refact: use react query on themeInfo
