+++ src/App.tsx
@@ -5,7 +5,13 @@ import AliveCheckPanel from "./apis/AliveCheckPanel"; // for testing
 import ToastContextProvider from "./contexts/ToastContextProdiver";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 
-const queryClient = new QueryClient();
+const queryClient = new QueryClient({
+  defaultOptions: {
+    queries: {
+      retry: false
+    }
+  }
+});
 
 function App() {
   return (
